Move chain-specific scripts to blockchain.
