+++ script/generic/update-all.ts
@@ -1,15 +1,15 @@
-import { BinanceAction } from "../action/binance";
-import { CosmosAction } from "../action/cosmos";
+import { BinanceAction } from "../blockchain/binance";
+import { CosmosAction } from "../blockchain/cosmos";
 import { EthForks } from "../generic/eth-forks";
 import { FoldersFiles } from "../generic/folders-and-files";
 import { JsonAction } from "../generic/json-format";
-import { KavaAction } from "../action/kava";
+import { KavaAction } from "../blockchain/kava";
 import { LogoSize } from "../generic/logo-size";
-import { TerraAction } from "../action/terra";
-import { TezosAction } from "../action/tezos";
-import { TronAction } from "../action/tron";
+import { TerraAction } from "../blockchain/terra";
+import { TezosAction } from "../blockchain/tezos";
+import { TronAction } from "../blockchain/tron";
 import { Validators } from "../generic/validators";
-import { WavesAction } from "../action/waves";
+import { WavesAction } from "../blockchain/waves";
 import { Allowlist } from "../generic/allowlists";
 import { ActionInterface, CheckStepInterface } from "../generic/interface";
 import * as chalk from 'chalk';
